job-search-with-claude/requirements.txt
# Core libraries
click # ==8.1.3
requests # ==2.26.0
PyYAML # ==6.0

# Natural Language Processing
spacy==3.5.0
# en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
numpy<2.0.0

# CLI Enhancements
rich # ==13.3.1

# Testing (optional, for future development)
pytest # ==7.2.1

# Type checking (optional, for development)
mypy # ==1.0.0
